LayoutTests/imported/w3c/web-platform-tests/html/cross-origin-embedder-policy/dedicated-worker.https-expected.txt

CONSOLE MESSAGE: Refused to load 'https://localhost:9443/html/cross-origin-embedder-policy/resources/dedicated-worker.js' worker because of Cross-Origin-Embedder-Policy.
CONSOLE MESSAGE: Worker load was blocked by Cross-Origin-Embedder-Policy
CONSOLE MESSAGE: Cannot load https://localhost:9443/html/cross-origin-embedder-policy/resources/dedicated-worker.js due to access control checks.
CONSOLE MESSAGE: Cancelled load to https://127.0.0.1:9443/common/blank.html because it violates the resource's Cross-Origin-Resource-Policy response header.
CONSOLE MESSAGE: Cancelled load to https://127.0.0.1:9443/common/blank.html because it violates the resource's Cross-Origin-Resource-Policy response header.
CONSOLE MESSAGE: Refused to load 'https://localhost:9443/html/cross-origin-embedder-policy/resources/dedicated-worker.js' worker because of Cross-Origin-Embedder-Policy.
CONSOLE MESSAGE: Worker load was blocked by Cross-Origin-Embedder-Policy
CONSOLE MESSAGE: Cannot load https://localhost:9443/html/cross-origin-embedder-policy/resources/dedicated-worker.js due to access control checks.
CONSOLE MESSAGE: Cancelled load to https://127.0.0.1:9443/common/blank.html because it violates the resource's Cross-Origin-Resource-Policy response header.
CONSOLE MESSAGE: Cancelled load to https://127.0.0.1:9443/common/blank.html because it violates the resource's Cross-Origin-Resource-Policy response header.

PASS COEP: none worker in COEP: none frame
PASS COEP: none worker in COEP: require-corp frame
PASS COEP: require-corp worker in COEP: none frame
PASS COEP: require-corp worker in COEP: require-corp frame
PASS COEP: none module worker in COEP: none frame
PASS COEP: none module worker in COEP: require-corp frame
PASS COEP: require-corp module worker in COEP: none frame
PASS COEP: require-corp module worker in COEP: require-corp frame
